Refactored code to use 1 timeline.js file
