+++ src/test/Timeline.test.js
@@ -1,6 +1,6 @@
 import {shallow} from 'enzyme';
 import React from 'react';
-import UserTweets from '../js/components/UserTweets';
+import Timeline from '../js/components/Timeline';
 
 const e = React.createElement;
 
@@ -11,7 +11,7 @@ let message = "button message",
     date = "December 17, 1995 03:24:00",
     id = "123456";
 
-/* Test UserTweets Component */
+/* Test HomeTimeline Component */
 const callBackFunc = (cb) => {
     cb(jsonObj);
 }
@@ -21,16 +21,46 @@ const json = (message, handle, name, img, created, id) => {
             '"user": {' +
                 '"twitterHandle":"' + handle + '",' +
                 '"name":"' + name + '",' +
-                '"profileImageURL":"' + img + '"},' +
+                '"profileImageURL":"' + img +
+            '"},' +
             '"createdAt":"' + created + '",' +
             '"id":"' + id + '"}]';
 }
 
 let jsonObj = JSON.parse(json(message, handle, name, img, date, id));
 
-describe('UserTweets Test No Input', () => {
+describe('Home Timeline Test No Input', () => {
     it('Should have children components but nothing else', () => {
-        const wrapper = shallow(e(UserTweets));
+        const wrapper = shallow(e(Timeline));
+        expect(wrapper.find('div').length).toEqual(3);
+        expect(wrapper.childAt(0).childAt(0).text()).toEqual("Home Timeline");
+        expect(wrapper.childAt(1).hasClass('buttonContainer1')).toEqual(true);
+        expect(wrapper.childAt(2).hasClass('dataHome')).toEqual(true);
+        expect(wrapper.childAt(2).children().length).toEqual(1);
+    });
+});
+
+describe('Home Timeline Test Valid Inputs and Button Press', () => {
+    it('Should have proper children components after button press', () => {
+        const wrapper = shallow(e(Timeline, {test: true, testFunc: callBackFunc}));
+        expect(wrapper.find('div').length).toEqual(3);
+        wrapper.find('button').simulate('click');
+        expect(wrapper.find('div').length).toEqual(6);
+
+        let itemElem = wrapper.childAt(2).childAt(0);
+        expect(itemElem.children().length).toEqual(2);
+        expect(itemElem.hasClass('item')).toEqual(true);
+        expect(itemElem.childAt(1).hasClass('message')).toEqual(true);
+
+        let messageElem = itemElem.childAt(1);
+        expect(messageElem.childAt(0).text()).toEqual('Dec 17');
+        expect(messageElem.childAt(1).text()).toEqual(message);
+    });
+});
+
+describe('User Timeline Test No Input', () => {
+    it('Should have children components but nothing else', () => {
+        const wrapper = shallow(e(Timeline, {displayUserTimeline: true}));
         expect(wrapper.find('div').length).toEqual(3);
         expect(wrapper.childAt(0).childAt(0).text()).toEqual("User Timeline");
         expect(wrapper.childAt(1).hasClass('buttonContainer2')).toEqual(true);
@@ -39,9 +69,9 @@ describe('UserTweets Test No Input', () => {
     });
 });
 
-describe('UserTweets Test Valid Inputs and Button Press', () => {
+describe('User Timeline Test Valid Inputs and Button Press', () => {
     it('Should have proper children components after button press', () => {
-        const wrapper = shallow(e(UserTweets, {test: true, testFunc: callBackFunc}));
+        const wrapper = shallow(e(Timeline, {displayUserTimeline: true, test: true, testFunc: callBackFunc}));
         expect(wrapper.find('div').length).toEqual(3);
         wrapper.find('button').simulate('click');
         expect(wrapper.find('div').length).toEqual(6);
@@ -49,13 +79,9 @@ describe('UserTweets Test Valid Inputs and Button Press', () => {
         let itemElem = wrapper.childAt(2).childAt(0);
         expect(itemElem.children().length).toEqual(2);
         expect(itemElem.hasClass('item')).toEqual(true);
-        expect(itemElem.childAt(0).hasClass('user')).toEqual(true);
         expect(itemElem.childAt(1).hasClass('message')).toEqual(true);
 
-        let userElem = itemElem.childAt(0);
         let messageElem = itemElem.childAt(1);
-        expect(userElem.childAt(1).text()).toEqual(name);
-        expect(userElem.childAt(2).text()).toEqual(handle);
         expect(messageElem.childAt(0).text()).toEqual('Dec 17');
         expect(messageElem.childAt(1).text()).toEqual(message);
     });
